Move BPF program loading to BpfProgram
Since we now have a class to represent a BPF program, the code for
loading belongs there. This will particularly allow us to better
decouple program loading from attachment and eventually delegate the
loading part to libbpf.

This also drops fd caching for BPF programs. The reason is that we're
moving program loading to libbpf so it doesn't make sense anyway. In
addition, progfd caching was only implemented for kprobes and we now
have kprobe_multi for that purpose.

Anyways, this still is a slight change in how wildcarded kprobes are
handled (especially when kprobe_multi is not present), so add new
runtime tests for them.
